Create MovieCollection.cs
built movie collection class, it will be used to associate a movie with one or more collections as well to order or prioritize movies within a collection. it will result in a many to many table or a linking table. its goal is to hold a bunch of primary key pairs. having a movie primary key and a collections primary key linking the two together.