Create IImageService.cs
Created an IImageService Interface which defines the minimum set of functionality when working with images in our application. The details of the service will be concerned with tranforming images into a data format suitable for storage into a database. as a consequnce with displaying these images we will have to convert them again into a format suitable for display.