Intro to the Seed Service.cs
Service Classes, introducing the seed service. The purpose of which is to add records to specific database tables so that a limited amount of information is available at the time of startup. The records that will be seeded will be a single user record, a single identity role and a single role assignment. basically making my own user record and making myself an administrator. in addition I will be seeding a default collection called 'all'.