Update MoviesController.cs
Added a library action to the already existing moviescontroller, this will allow a user to make  a request to the homecontroller and library action, which will display all the imported movies uncategorized and outside of their respective movies collection. This same action can be used by both users and administrators, with a difference being what they can do once they are on the page.