Update MoviesController.cs
created the movies controller details action, this allows us to go to the details page of the movie and that movie may be coming from the api and it may be coming from the database. In order to manage the two sources of movie data, using a boolean flag to indicate where the detail data is coming from.